@@ -1,7 +1,7 @@
 import { APP_INITIALIZER, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { HttpClientModule } from '@angular/common/http';
+import { HttpClient, HttpClientModule } from '@angular/common/http';
 
 /* MAIN COMPONENT */
 import { AppComponent } from './app.component';
@@ -23,14 +23,15 @@ import { ConfigService, configFactory } from "../config";
 /* Third Party */
 import { ModalModule } from 'ngx-bootstrap/modal';
 import { JwtModule } from '@auth0/angular-jwt';
-
+import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 
 // settings for JWT
 export function tokenGetter(): string {
 	return localStorage.getItem('id_token') || "";
 }
 
-var jwtOptions = {
+const jwtOptions = {
 	config: {
 		tokenGetter: tokenGetter,
 		whitelistedDomains: environment.jwtDomains,
@@ -39,32 +40,41 @@ var jwtOptions = {
 	}
 };
 
+export function createTranslateLoader(http: HttpClient) {
+	return new TranslateHttpLoader(http, './assets/text/', '.json');
+}
+
 @NgModule({
 	imports: [
-	BrowserModule,
+		BrowserModule,
 		BrowserAnimationsModule,
 		HttpClientModule,
-		
 		AppRoutingModule,
-		
 		SharedModule,
 		LoginModule,
 		NotFoundPageModule,
-
 		ModalModule.forRoot(),
-		JwtModule.forRoot(jwtOptions)
+		JwtModule.forRoot(jwtOptions),
+		TranslateModule.forRoot({
+			defaultLanguage: 'cs',
+			loader: {
+				provide: TranslateLoader,
+				useFactory: createTranslateLoader,
+				deps: [HttpClient],
+			}
+		}),
 	],
 	declarations: [
-		AppComponent		
+		AppComponent
 	],
 	entryComponents: [
 	],
-	providers: [		
-		{ 
-			provide: APP_INITIALIZER, 
+	providers: [
+		{
+			provide: APP_INITIALIZER,
 			useFactory: configFactory,
 			deps: [ConfigService],
-			multi: true 
+			multi: true
 		},
 		httpInterceptorProviders
 	],
